src/requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
shap
pyspark
prometheus-client
requests
imbalanced-learn
lime
kafka-python
